requirements.txt
pyautogui
pygetwindow
opencv-python
numpy
Pillow